refactor: cut over projectiles into package modules
