+++ src/crimson/projectiles/__init__.py
@@ -1,0 +1,45 @@
+from __future__ import annotations
+
+from .pools import (
+    PROJECTILE_BEHAVIOR_BY_TYPE_ID,
+    ProjectileBehavior,
+    ProjectilePool,
+    SecondaryProjectilePool,
+    _within_native_find_radius,
+)
+from .types import (
+    CreatureDamageApplier,
+    Damageable,
+    FxQueueLike,
+    MAIN_PROJECTILE_POOL_SIZE,
+    PlayerDamageable,
+    Projectile,
+    ProjectileHit,
+    ProjectileRuntimeState,
+    ProjectileTypeId,
+    SecondaryDetonationKillHandler,
+    SECONDARY_PROJECTILE_POOL_SIZE,
+    SecondaryProjectile,
+    SecondaryProjectileTypeId,
+)
+
+__all__ = [
+    "CreatureDamageApplier",
+    "Damageable",
+    "FxQueueLike",
+    "MAIN_PROJECTILE_POOL_SIZE",
+    "PlayerDamageable",
+    "PROJECTILE_BEHAVIOR_BY_TYPE_ID",
+    "Projectile",
+    "ProjectileBehavior",
+    "ProjectileHit",
+    "ProjectilePool",
+    "ProjectileRuntimeState",
+    "ProjectileTypeId",
+    "SECONDARY_PROJECTILE_POOL_SIZE",
+    "SecondaryDetonationKillHandler",
+    "SecondaryProjectile",
+    "SecondaryProjectilePool",
+    "SecondaryProjectileTypeId",
+    "_within_native_find_radius",
+]
